+++ test_run.py
@@ -24,7 +24,7 @@ def run(send: bool, limit: int = None):
         arxiv_query=cfg["arxiv"]["query"],
         max_results=int(cfg["arxiv"].get("max_results", 30)),
         only_new=bool(cfg["arxiv"].get("only_new", True)),
-        days_back=int(cfg["arxiv"].get("days_back", 1)),
+        days_back=float(cfg["arxiv"].get("days_back", 1)),
         source=str(cfg["arxiv"].get("source", "rss")).lower(),
     )
